Changed to full screen, reverted x-axis rotation
Changed to full screen for our demo coming up on Wednesday. Also reverted x-axis (up/down) rotation changes because there is no point in looking up or down—easier to control now.

Not committed: I tried fixing the camera so that it would be more "first-person" cause technically we're still using a 3rd person camera, but the clipping of the walls when the player gets too close gets worse with the "first-person" camera.